requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiohttp==3.8.6
    # via discord-py
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.3
    # via aiohttp
attrs==23.1.0
    # via aiohttp
build==1.0.3
    # via pip-tools
certifi==2023.7.22
    # via requests
charset-normalizer==3.3.0
    # via
    #   aiohttp
    #   requests
click==8.1.7
    # via pip-tools
discord-py==2.3.2
    # via -r requirements.in
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.4
    # via
    #   requests
    #   yarl
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
packaging==23.2
    # via build
pip-tools==7.3.0
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via build
python-dotenv==1.0.0
    # via -r requirements.in
requests==2.31.0
    # via -r requirements.in
urllib3==2.0.7
    # via requests
wheel==0.41.2
    # via pip-tools
yarl==1.9.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
